Change sc_corr implementation to remove unnecessary calc
Using elementwise operations, and JIT compilation, with help of Numba for numpy, and cp.Elementwise kernel for Cupy.
This helps returning early in case of numerator being zero, which helps with not requiring to divide with a conditional mask (which we had to do to prevent NaNs)
Also prevents calculation of two denom arrays.

Lastly, find_corr_mat return uint32 values